+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for `At Home` dates
 - Add support for non-Instagram date images, i.e. full image urls
 - Hide neighborhoods with no dates
+- Sort tags alphabetically
+- Update "Discover Dates" row to feature "At Home" dates
 
 ## [v2.0.0](https://github.com/nathanab/beacon/compare/v1.2.0...v2.0.0)
 - Refactor app to use server-side rendering via next.js